Arduinoのモデルによってintのサイズが異なるのでlongにして対応

+++ ekispert/part1/arduino-ekispert-part1.ino
@@ -5,7 +5,7 @@
 // 変数の定義　Definition of variables
 SakuraIO_I2C sakuraio;
 const int B = 4250;               // B value of the thermistor
-const int R0 = 100000;            // R0 = 100k
+const long R0 = 100000;           // R0 = 100k
 const int pinTempSensor = A0;     // Grove - Temperature Sensor connect to A0
 uint32_t cnt = 0;
